tools/requirements.txt
azure-kusto-data>=6.0.0
azure-identity>=1.25.0
mcp>=1.27.0
